+++ CHANGES.rst
@@ -22,6 +22,7 @@ Changelog
 * [mpm] Automate publishing of package to PyPi on tagging.
 * [mpm] Save build artifacts on each CI runs.
 * [mpm] Automate auto-sorting of module imports.
+* [mpm] Auto-fix common typos.
 * [mpm] Automate GitHub label generation and synchronization.
 * [mpm] Automaticcaly applies labels on PRs and issues depending on their
   changed files and content.